Using SQL with JPA to solve some querys and makes the data flow easier

+++ wolox/src/main/java/com/tech/wolox/dto/AlbumDTO.java
@@ -19,7 +19,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AlbumDTO {
-    private Long userId;
-    private Long id;
+    private Integer userId;
+    private Integer id;
     private String title;
 }
